Refactor `oe_gen_marshal_struct`
This refactors the previously borrowed `mk_ms_member_decl` function from
Intel into a similar member emitter specific to our use-case, and
refactors `oe_gen_marshal_struct` to take a `Trusted` or `Untrusted`
function and emit the marshalling struct. The code has been written to
all exist in one place (so it can be read more easily) instead of spread
over several functions that were only used for this single bit of logic.

The newline generation was corrected as well, resolving some outstanding
TODOs.

There remains a TODO to determine how we can remove our requirement of
the strlen variable for ocalls, as Intel does not require this (hence
the previous note in the borrowed code), but we still need it.