Clean up by using |> and polymorphic []
Instead of adding lists of the empty string, `[""]`, because `[]` is
polymorphic, it just doesn't add the element at all. This removes the
need to remove empty strings later.

Also, by using the `|>` operator, we can move boring calls to
`List.flatten` to the end of a chain instead of the beginning.

Finally, get rid of pointless single argument lambdas with the
expression itself is a single argument function.