Expose deep-copy prototype only through `--experimental` flag
The edger8r must be called with `--experimental` to take advantage of
the prototyped deep-copy feature. This should allow us to merge upstream
more easily.

Note that both the enclave and host side must use the experimental flag.
In the future, we should somehow create a more clear compilation failure
if trusted and untrusted code is combined with mismatched feature
support.