Fix handling of `in, out` pointers for deep copy
This was a very strange bug. When we generate the pointer setters, we
actually need to default to `SET_IN_OUT` (and `COPY_SET_IN_OUT`) in our
predicates. That is, when recursively examining nested pointers, they no
longer satisfy `is_in_ptr` nor `is_out_ptr` nor `is_inout_ptr` because
they (presumably) have no direction.

We need to circle back and verify if their direction is indeed
`PtrNoDirection`. For now, we're going to play it "safe" in order to
enable the prototype by pretending they're `PtrInOut`; however, we will
need to pass the direction of the top-level struct that is being
deep-copied in order to get this correct and not accidentally marshal
sensitive information out of the enclave if it was only meant to go in.