Fix save/restore logic for out pointers
When unmarshalling out pointers of nested structs, we need to restore
the original pointer values because first we overwrite the structs on
the host side when unmarshalling the struct.

Annoyingly, this means we need to save each out (or in/out) pointer
value on the host side. In C++ I'd use a vector here, but after working
with our `queue.h` implementation, I had to pivot to a limited
implementation that cannot handle more than 512 pointers. We should be
able to handle N pointers, but I need a performant, extendable data
structure to hold all the saved pointers (a FIFO queue would be nice).

A couple commits ago we incorrectly attempted to use the values stored
in `_args` to cheaply save and restore. This looked fine at first, but
the macros `OE_WRITE...` update the values in `_args` to point into the
input buffer. This should not have worked at all, as we were just
pointing to random memory.

Finally, it is not necessary for us to recursively assign the initial
arguments to `_args`, so instead we fix that logic, and then follow it
with the pointer saving logic (which is similar, but saves into the
`void* ptrs[]` array.)

Also, our null checks are checking `foo && foo[i].bar` but notably not
checking `foo && foo[i] && foo[i].bar`, which needs to be fixed (see
TODOs).