Run ocamlformat 0.11 on edger8r
With esy we have access to an up-to-date (and pinned) version of
`ocamlformat`, which also formats our `dune` files. This newer version
also allows us to include only our code, excluding the Intel code (and
the copy of `main.ml`).

Manually, this can be run via `esy dune build @fmt --auto-promote`, but
this will also be integrated into the build system.